requirements.txt
Django==2.2.5
django-import-export==1.2.0
djangorestframework==3.10.3
markdown==3.1.1
django-filter==2.2.0
djangorestframework-simplejwt==4.3.0
django-rest-auth==0.9.5
psycopg2==2.8.3
gunicorn==19.9.0
django-heroku==0.3.1